requirements.txt
astroid==2.8.6
autopep8==1.6.0
flake8==4.0.1
infi.systray==0.1.12
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mido==1.2.10
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.11.1
python-osc==1.8.0
python-rtmidi==1.4.9
PyYAML==6.0
toml==0.10.2
typing_extensions==4.0.0
wrapt==1.13.3
